Add definitions of built-in translation functions (#9)
NVDA installs several functions used for translations to the global built-in namespace. Since they're added into globals VS Code is not aware about them which results in warnings in the problems view. There is also no Intellisense for these functions. This Pr ads type stub with definitions of the translation functions, and ensures it is used by PyLance.