Add new python.analysis.extraPaths setting for Pylance (#10)
When the default language server for the VS Code Python extension was switched from Microsoft Python Language Server to Pylance some settings were renamed. While this migration occurs automatically it causes VS Code config to be marked as dirty in Git. In this PR I have added a new setting python.analysis.extraPaths which is equivalent to the older python.autoComplete.extraPaths. Note that since it is possible to use a different (notably Jedi) language server for Python in VS Code, and I failed to find any conclusive documentation as to which setting it uses I've kept the old python.autoComplete.extraPaths setting as its presence causes no harm and may be needed for users of different language servers.